+++ Watt/Utilities/CheckedContinuationReference.swift
