+++ Watt/Watt-Bridging-Header.h
@@ -2,5 +2,5 @@
 //  Use this file to import your target's public headers that you would like to expose to Swift.
 //
 
-#import "SimpleProxy.h"
+#import "Utilities/SimpleProxy.h"
 #import "Extensions/NSObject+Extensions.h"
